requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.9

Werkzeug==2.0.3
